move all to __init__ again
